reorganize type domain
