+++ src/crab/dsl_syntax.hpp
@@ -33,6 +33,16 @@ inline linear_constraint_t operator>=(const linear_expression_t& e1, const linea
 
 inline linear_constraint_t operator>(const linear_expression_t& e1, const linear_expression_t& e2) { return e2 < e1; }
 
+inline linear_constraint_t eq(const variable_t a, const variable_t b) {
+    using namespace crab::dsl_syntax;
+    return {a - b, constraint_kind_t::EQUALS_ZERO};
+}
+
+inline linear_constraint_t neq(const variable_t a, const variable_t b) {
+    using namespace crab::dsl_syntax;
+    return {a - b, constraint_kind_t::NOT_ZERO};
+}
+
 inline linear_constraint_t operator==(const linear_expression_t& e1, const linear_expression_t& e2) {
     return linear_constraint_t(e1 - e2, constraint_kind_t::EQUALS_ZERO);
 }
