+++ src/crab/split_dbm.cpp
@@ -6,6 +6,7 @@
 #include "crab_utils/debug.hpp"
 #include "crab_utils/stats.hpp"
 #include "string_constraints.hpp"
+#include "type_encoding.hpp"
 
 namespace crab::domains {
 
@@ -479,14 +480,14 @@ SplitDBM SplitDBM::operator|(const SplitDBM& o) const& {
 
     // Build the permuted view of x and y.
     assert(g.size() > 0);
-    GraphPerm<const graph_t> gx(perm_x, g);
+    GraphPerm gx(perm_x, g);
     assert(o.g.size() > 0);
-    GraphPerm<const graph_t> gy(perm_y, o.g);
+    GraphPerm gy(perm_y, o.g);
 
     // Compute the deferred relations
     graph_t g_ix_ry;
     g_ix_ry.growTo(sz);
-    SubGraph<GraphPerm<const graph_t>> gy_excl(gy, 0);
+    SubGraph gy_excl(gy, 0);
     for (vert_id s : gy_excl.verts()) {
         for (vert_id d : gy_excl.succs(s)) {
             if (auto ws = gx.lookup(s, 0)) {
@@ -500,13 +501,13 @@ SplitDBM SplitDBM::operator|(const SplitDBM& o) const& {
     bool is_closed;
     graph_t g_rx(GrOps::meet(gx, g_ix_ry, is_closed));
     if (!is_closed) {
-        GrOps::apply_delta(g_rx, GrOps::close_after_meet(SubGraph<graph_t>(g_rx, 0), pot_rx, gx, g_ix_ry));
+        GrOps::apply_delta(g_rx, GrOps::close_after_meet(SubGraph(g_rx, 0), pot_rx, gx, g_ix_ry));
     }
 
     graph_t g_rx_iy;
     g_rx_iy.growTo(sz);
 
-    SubGraph<GraphPerm<const graph_t>> gx_excl(gx, 0);
+    SubGraph gx_excl(gx, 0);
     for (vert_id s : gx_excl.verts()) {
         for (vert_id d : gx_excl.succs(s)) {
             // Assumption: gx.mem(s, d) -> gx.edge_val(s, d) <= ranges[var(s)].ub() - ranges[var(d)].lb()
@@ -521,7 +522,7 @@ SplitDBM SplitDBM::operator|(const SplitDBM& o) const& {
     // Similarly, should use a SubGraph view.
     graph_t g_ry(GrOps::meet(gy, g_rx_iy, is_closed));
     if (!is_closed) {
-        GrOps::apply_delta(g_ry, GrOps::close_after_meet(SubGraph<graph_t>(g_ry, 0), pot_ry, gy, g_rx_iy));
+        GrOps::apply_delta(g_ry, GrOps::close_after_meet(SubGraph(g_ry, 0), pot_ry, gy, g_rx_iy));
     }
 
     // We now have the relevant set of relations. Because g_rx and g_ry are closed,
@@ -637,9 +638,9 @@ SplitDBM SplitDBM::widen(const SplitDBM& o) const {
 
     // Build the permuted view of x and y.
     assert(g.size() > 0);
-    GraphPerm<const graph_t> gx(perm_x, g);
+    GraphPerm gx(perm_x, g);
     assert(o.g.size() > 0);
-    GraphPerm<const graph_t> gy(perm_y, o.g);
+    GraphPerm gy(perm_y, o.g);
 
     // Now perform the widening
     vert_set_t widen_unstable(unstable);
@@ -709,9 +710,9 @@ std::optional<SplitDBM> SplitDBM::meet(const SplitDBM& o) const {
 
     // Build the permuted view of x and y.
     assert(g.size() > 0);
-    GraphPerm<const graph_t> gx(perm_x, g);
+    GraphPerm gx(perm_x, g);
     assert(o.g.size() > 0);
-    GraphPerm<const graph_t> gy(perm_y, o.g);
+    GraphPerm gy(perm_y, o.g);
 
     // Compute the syntactic meet of the permuted graphs.
     bool is_closed;
@@ -726,7 +727,7 @@ std::optional<SplitDBM> SplitDBM::meet(const SplitDBM& o) const {
     }
 
     if (!is_closed) {
-        GrOps::apply_delta(meet_g, GrOps::close_after_meet(SubGraph<graph_t>(meet_g, 0), meet_pi, gx, gy));
+        GrOps::apply_delta(meet_g, GrOps::close_after_meet(SubGraph(meet_g, 0), meet_pi, gx, gy));
 
         // Recover updated LBs and UBs.<
 
@@ -888,7 +889,7 @@ void SplitDBM::assign(variable_t lhs, const linear_expression_t& e) {
         // apply_delta should be safe here, as x has no edges in G.
         GrOps::apply_delta(g, delta);
     }
-    GrOps::apply_delta(g, GrOps::close_after_assign(SubGraph<graph_t>(g, 0), potential, vert));
+    GrOps::apply_delta(g, GrOps::close_after_assign(SubGraph(g, 0), potential, vert));
 
     if (lb_w) {
         g.update_edge(vert, *lb_w, 0);
@@ -937,7 +938,7 @@ void SplitDBM::normalize() {
     edge_vector delta;
     // GrOps::close_after_widen(g, potential, vert_set_wrap_t(unstable), delta);
     // GKG: Check
-    GrOps::apply_delta(g, GrOps::close_after_widen(SubGraph<graph_t>(g, 0), potential, vert_set_wrap_t(unstable)));
+    GrOps::apply_delta(g, GrOps::close_after_widen(SubGraph(g, 0), potential, vert_set_wrap_t(unstable)));
     // Retrieve variable bounds
     GrOps::apply_delta(g, GrOps::close_after_assign(g, potential, 0));
 
@@ -1087,9 +1088,6 @@ static std::string to_string(variable_t vd, variable_t vs, const SplitDBM::Weigh
     return elem.str();
 }
 
-static const std::vector<std::string> type_string = {"shared", "stack",  "packet",          "ctx",
-                                                     "number", "map_fd", "map_fd_programs", "uninitialized"};
-
 string_invariant SplitDBM::to_set() const {
     if (this->is_top()) {
         return string_invariant::top();
@@ -1099,8 +1097,8 @@ string_invariant SplitDBM::to_set() const {
     // Intervals
 
     // Extract all the edges
-    SubGraph<const SplitDBM::graph_t> g_excl(this->g, 0);
-    for (SplitDBM::vert_id v : g_excl.verts()) {
+    SubGraph g_excl(this->g, 0);
+    for (vert_id v : g_excl.verts()) {
         if (!this->rev_map[v]) {
             continue;
         }
@@ -1117,21 +1115,21 @@ string_invariant SplitDBM::to_set() const {
         std::stringstream elem;
         elem << variable;
         if (variable.is_type()) {
-            int lb = (int)v_out.lb().number().value();
-            int ub = (int)v_out.ub().number().value();
+            auto lb = static_cast<type_encoding_t>(v_out.lb().number().value());
+            auto ub = static_cast<type_encoding_t>(v_out.ub().number().value());
             if (lb == ub) {
                 if (variable.is_in_stack() && lb == T_NUM) {
                     // no need to show this
                     continue;
                 }
-                elem << "=" << type_string.at(-lb);
+                elem << "=" << type_encoding_to_string(lb);
             } else {
                 elem << " in {";
-                for (int type = lb; type <= ub; type++) {
+                for (auto type : iterate_types(lb, ub)) {
                     if (type > lb) {
                         elem << ", ";
                     }
-                    elem << type_string.at(-type);
+                    elem << type_encoding_to_string(type);
                 }
                 elem << "}";
             }
@@ -1147,12 +1145,12 @@ string_invariant SplitDBM::to_set() const {
     }
 
     std::set<std::tuple<variable_t, variable_t, Weight>> diff_csts;
-    for (SplitDBM::vert_id s : g_excl.verts()) {
+    for (vert_id s : g_excl.verts()) {
         if (!this->rev_map[s]) {
             continue;
         }
         variable_t vs = *this->rev_map[s];
-        for (SplitDBM::vert_id d : g_excl.succs(s)) {
+        for (vert_id d : g_excl.succs(s)) {
             if (!this->rev_map[d]) {
                 continue;
             }
@@ -1237,7 +1235,7 @@ bool SplitDBM::entail(const linear_constraint_t& rhs) const {
     interval_t interval = eval_interval(rhs.expression());
     switch (rhs.kind()) {
     case constraint_kind_t::EQUALS_ZERO:
-        if (interval.singleton() == std::optional<number_t>(number_t(0))) {
+        if (interval.singleton() == std::optional(number_t(0))) {
             return true;
         }
         break;
@@ -1305,9 +1303,9 @@ static interval_t get_interval(const SplitDBM::vert_map_t& m, const SplitDBM::gr
 }
 
 interval_t SplitDBM::get_interval(variable_t x, int finite_width) const {
-    return crab::domains::get_interval(vert_map, g, x, finite_width);
+    return domains::get_interval(vert_map, g, x, finite_width);
 }
 
-interval_t SplitDBM::operator[](variable_t x) const { return crab::domains::get_interval(vert_map, g, x, 0); }
+interval_t SplitDBM::operator[](variable_t x) const { return domains::get_interval(vert_map, g, x, 0); }
 
 } // namespace crab::domains
