+++ src/crab_utils/num_big.hpp
@@ -39,6 +39,11 @@ class number_t final {
     }
 
     template <is_enum T>
+    explicit operator T() const {
+        return static_cast<T>(_n);
+    }
+
+    template <is_enum T>
     T cast_to() const {
         return static_cast<T>(static_cast<std::underlying_type_t<T>>(_n));
     }
